feat: first push

+++ database/migrations/2022_03_18_061422_create_products_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('brand');
             $table->string('category');
             $table->string('color');
-            $table->decimal('price', 7, 2);
+            $table->integer('price');
             $table->softDeletes();
             $table->timestamps();
         });
